pip_requirements.txt
cffi
jupyterlab==4.3.5
numpy==1.26.4
uproot==5.5.1
matplotlib
scipy
pandas==2.2.2
awkward-pandas
nbstripout
tqdm
ipywidgets
dill
h5py
tables
emcee
corner
lz4
xxhash
fsspec-xrootd
landaupy
